Added items and multi-item recipies
An item (currently) is just an id. every buffer has one id attached, and
every machine has some new fields. requires is the list of items and
their quantities needed for the recipe, and creates is the list of items
and quantities that come out. Each element in each of those lists is
given a buffer